requirements.txt
streamlit
PyMuPDF
Pillow
requests
python-dotenv
watchdog
beautifulsoup4
